requirements.txt
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
psycopg2==2.8.6
python-dotenv==0.15.0